+++ python/instrumentation/openinference-instrumentation-mcp/src/openinference/instrumentation/mcp/__init__.py
@@ -21,25 +21,43 @@ class MCPInstrumentor(BaseInstrumentor):  # type: ignore
     def _instrument(self, **kwargs: Any) -> None:
         register_post_import_hook(
             lambda _: wrap_function_wrapper(
-                "mcp.client.sse", "sse_client", self._transport_wrapper
+                "mcp.client.streamable_http",
+                "streamablehttp_client",
+                self._wrap_transport_with_callback,
+            ),
+            "mcp.client.streamable_http",
+        )
+
+        register_post_import_hook(
+            lambda _: wrap_function_wrapper(
+                "mcp.server.streamable_http",
+                "StreamableHTTPServerTransport.connect",
+                self._wrap_plain_transport,
+            ),
+            "mcp.server.streamable_http",
+        )
+
+        register_post_import_hook(
+            lambda _: wrap_function_wrapper(
+                "mcp.client.sse", "sse_client", self._wrap_plain_transport
             ),
             "mcp.client.sse",
         )
         register_post_import_hook(
             lambda _: wrap_function_wrapper(
-                "mcp.server.sse", "SseServerTransport.connect_sse", self._transport_wrapper
+                "mcp.server.sse", "SseServerTransport.connect_sse", self._wrap_plain_transport
             ),
             "mcp.server.sse",
         )
         register_post_import_hook(
             lambda _: wrap_function_wrapper(
-                "mcp.client.stdio", "stdio_client", self._transport_wrapper
+                "mcp.client.stdio", "stdio_client", self._wrap_plain_transport
             ),
             "mcp.client.stdio",
         )
         register_post_import_hook(
             lambda _: wrap_function_wrapper(
-                "mcp.server.stdio", "stdio_server", self._transport_wrapper
+                "mcp.server.stdio", "stdio_server", self._wrap_plain_transport
             ),
             "mcp.server.stdio",
         )
@@ -63,7 +81,18 @@ class MCPInstrumentor(BaseInstrumentor):  # type: ignore
         unwrap("mcp.server.stdio", "stdio_server")
 
     @asynccontextmanager
-    async def _transport_wrapper(
+    async def _wrap_transport_with_callback(
+        self, wrapped: Callable[..., Any], instance: Any, args: Any, kwargs: Any
+    ) -> AsyncGenerator[Tuple["InstrumentedStreamReader", "InstrumentedStreamWriter", Any], None]:
+        async with wrapped(*args, **kwargs) as (read_stream, write_stream, get_session_id_callback):
+            yield (
+                InstrumentedStreamReader(read_stream),
+                InstrumentedStreamWriter(write_stream),
+                get_session_id_callback,
+            )
+
+    @asynccontextmanager
+    async def _wrap_plain_transport(
         self, wrapped: Callable[..., Any], instance: Any, args: Any, kwargs: Any
     ) -> AsyncGenerator[Tuple["InstrumentedStreamReader", "InstrumentedStreamWriter"], None]:
         async with wrapped(*args, **kwargs) as (read_stream, write_stream):
@@ -91,10 +120,12 @@ class InstrumentedStreamReader(ObjectProxy):  # type: ignore
         return await self.__wrapped__.__aexit__(exc_type, exc_value, traceback)
 
     async def __aiter__(self) -> AsyncGenerator[Any, None]:
-        from mcp.types import JSONRPCMessage, JSONRPCRequest
+        from mcp.shared.message import SessionMessage
+        from mcp.types import JSONRPCRequest
 
         async for item in self.__wrapped__:
-            request = cast(JSONRPCMessage, item).root
+            session_message = cast(SessionMessage, item)
+            request = session_message.message.root
 
             if not isinstance(request, JSONRPCRequest):
                 yield item
@@ -122,9 +153,11 @@ class InstrumentedStreamWriter(ObjectProxy):  # type: ignore
         return await self.__wrapped__.__aexit__(exc_type, exc_value, traceback)
 
     async def send(self, item: Any) -> Any:
-        from mcp.types import JSONRPCMessage, JSONRPCRequest
+        from mcp.shared.message import SessionMessage
+        from mcp.types import JSONRPCRequest
 
-        request = cast(JSONRPCMessage, item).root
+        session_message = cast(SessionMessage, item)
+        request = session_message.message.root
         if not isinstance(request, JSONRPCRequest):
             return await self.__wrapped__.send(item)
         meta = None