+++ python/instrumentation/openinference-instrumentation-agno/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.3](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-agno-v0.1.2...python-openinference-instrumentation-agno-v0.1.3) (2025-05-13)
+
+
+### Bug Fixes
+
+* **agno:** Allow agent.name to be `None` in `arun`. ([#1621](https://github.com/Arize-ai/openinference/issues/1621)) ([5288979](https://github.com/Arize-ai/openinference/commit/52889796daa5ac221af21b76d77b80e594cdcf89))
+
 ## [0.1.2](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-agno-v0.1.1...python-openinference-instrumentation-agno-v0.1.2) (2025-05-13)